refactor: add withdraw stubs
