+++ packages/sbtc/src/transactions/index.ts
@@ -1,3 +1,7 @@
 export * from './api';
-export * from './deposit';
 export * from './constants';
+
+export * from './deposit';
+export * from './withdraw';
+
+export * from './utils';
